+++ Malmo/samples/malmoenv/missions/mobchase_single_agent.xml
