+++ Malmo/samples/malmoenv/missions/mobchase_two_agents.xml
