+++ Malmo/samples/malmoenv/package.sh
@@ -1,0 +1,1 @@
+python3 setup.py sdist bdist_wheel
